Removed GCC compiler warning.

+++ Device/ARM/ARMSC300/Source/GCC/startup_ARMSC300.c
@@ -3,7 +3,7 @@
  * @brief    CMSIS Core Device Startup File for
  *           ARMSC300 Device Series
  * @version  V5.00
- * @date     02. March 2016
+ * @date     26. Aprl 2016
  ******************************************************************************/
 /*
  * Copyright (c) 2009-2016 ARM Limited. All rights reserved.
@@ -128,7 +128,7 @@ void SPI_IRQHandler      (void) __attribute__ ((weak, alias("Default_Handler")))
  *----------------------------------------------------------------------------*/
 const pFunc __Vectors[] __attribute__ ((section(".vectors"))) = {
   /* Cortex-M3 Exceptions Handler */
-  (pFunc)&__StackTop,                       /*      Initial Stack Pointer     */
+  (pFunc)((uint32_t)&__StackTop),           /*      Initial Stack Pointer     */
   Reset_Handler,                            /*      Reset Handler             */
   NMI_Handler,                              /*      NMI Handler               */
   HardFault_Handler,                        /*      Hard Fault Handler        */
